Revert audio focus logic in Telecom.
As part of the effort to resolve hidden dependencies, we switched to use
AudioManager#requestAudioFocus instead of the hidden API.  It turns out
the old hidden API did something special; it "locked" focus onto Telecom.
As a consequence another app could steal call focus from Telecom.

To fix this for now we'll use the old hidden APIs; there is a way to do
this was using AudioFocusRequest.Builder.setLocksFocus(true), however
that appears not to work due to a permissions error. Until this can be
resolved, we'll switch back to the old hidden API until we can figure out
the right workaround.

Added new CUJ tests to validate this behavior and ensure no regression
in the future.

Test: Added new CUJ tests to cover this case.
Bug: 374309755
Flag: com.android.server.telecom.flags.telecom_resolve_hidden_dependenciesNONE bugfix.
Change-Id: Ife53e94aa4b41432e736545d183487c5a715a3aa