Refrain from unbinding BT ICS when call is ongoing
Due to how we delay the unbinding from the BT ICS in
InCallController#onDisconnectedTonePlaying, it's possible that we will
end up unbinding when another call has started. If the call is
placed/received on the same phone account as the previous call, then we
will prematurely unbind from the BT ICS which will leave Telecom in a
state where we're unable to send call state updates to the BT stack. As
a result, BT stack will not know when this new call is disconnected
which causes the SCO stream to be persistent (causes audio to be
inaudible over HFP for non-call related use cases).

Bug: 370820369
Test: Manual to verify that we don't unbind from BT ICS when subsequent
calls are placed immediately after
Flag: EXEMPT bugfix

Change-Id: I107584b8a27ba2c48d6d87fb32bcc2dfc7e393b5