Resolve audio not routing to watch issue
When a BTHS and watch are paired with the phone, the call audio was
routing to BTHS instead of the watch when the call was answered on the
watch. Although the resolveActiveBtRoutingAndBtTimingIssue flag was
disabled when this was tested, the issue would still exist in the
flagged changes as well. Revert the changes where we wait to update the
active device once the CallAudioRouteController has finished processing
the requests and update the state immediately.

We should also ensure that when considering the watch device for
routing, also make sure to check it against the active device cache
instead of just considering the current route (it's possible that the
watch hasn't been routed into yet but the active device is the watch).

Another fix involves making sure that we ignore recalculating the
baseline in cases where the pending route or current route isn't already
routed into the BT device that should be excluded.

Bug: 377628039
Test: Manual with BTHS + watch paired and verifying audio routed to
watch when call answered on the watch.
Flag:
com.android.server.telecom.flags.resolve_active_bt_routing_and_bt_timing_issue

Change-Id: I47449430784efb1eda291b58d40dabd5fb8dd125