Resolve call audio focus not abandoned
If a BT device is connected and an incoming call is placed and rejected,
if the inband ringing is disabled, the audio routing never goes active.
The logic in handling a focus switch to NO_FOCUS currently only abandons
audio focus if the audio routing was active. Looking back at this code,
we really should just reinitialize all of the audio routing state
regardless of whether or not the audio routing was active. A lot of
these operations would be unaffected when performing them over an
inactive audio route but it's good to just reset the controller state
once the call ends.

Bug: 376820227
Flag: EXEMPT bugfix
Test: CallAudioRouteControllerTest#testAbandonCallAudioFocusAfterCallEnd
Change-Id: I3657cb8797d1a15eecf7254b9ceec9ea3851fde4