DSDA: Call sequencing support implementation
This CL adds the implementation details required to support call
sequencing in Telecom. The transactional side of this has largely been
unchanged in exception to cases where we don't wait for call state
changes (i.e. MaybeHoldCallForNewCallTransaction).

All of the call sequencing related implementation details can be found
in CallSequencingController. All of the logic in CallsManager has been
delegated to CallsManagerCallSequencingAdapter to determine which code
path to use depending on if sequencing is enabled in exception to some
cases where it didn't seem necessary to create redundant functions for
the old/new paths.

Bug: 327038818
Test: m Telecom (formal tests have not been included into this CL)
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: I2ebc80f7e7e87c3677f0e0953420a0a977e8031e