Fix NPE in BluetoothDeviceManager
The NPE was caused by a null check on mBluetoothAdapter in the constructor which skipped over the instantiation of mBluetoothHeadsetFuture. We should move the initialization of fields that aren't dependent on the adapter outside of the if block so that we don't run into other potential NPEs.

Bug: 380284503
Change-Id: I9a18630202af3ae4b329394d0cf688af71b90749
Flag: EXEMPT bugfix
Test: m Telecom