Create CallAudioWatchdog.
To aid in debugging voip app audio issues, add a CallAudioWatchdog which
watches the audio record and playback resources that voip apps allocate
and includes information about this in the Telecom dumpsys.
Audio sessions for an app that does not use the Telecom framework are
also logged as non-telecom calls in the dumpsys to further aid in
debugging issues reported by users when apps don't use Telecom.

Flag: com.android.server.telecom.flags.enable_call_audio_watchdog
Test: Added CallAudioWatchDog test.
Test: Performed manual testing with telecom aware and non-aware voip apps
to verify logging.
Bug: 384570270

Change-Id: I8b4c21ee7ddd0585115e67429870b4059660b8ce