handle 2 calls from the same app case in VoipCallMonitor
It was pointed out that AM was throwing an exception when a second
voip call was added to telecom for the same app.  Upon investigation,
foreground service delegation should only be granted once for an app
instead of on a per call basis.

Changes:
- FGS is started ONLY if the application does not have it already
  instead of every time the call is added
- FGS is revoked if the notification is removed and the call is
  not disconnected after 5 seconds
- only support call-style notifications posted AFTER the call is
  added
- The logging around FGS events has changed.  Prev logged when
  start/stop requests are sent to AM. Now, fgs gain/revoked is
  logged when AM calls the service.

Flag: com.android.server.telecom.flags.voip_call_monitor_refactor
Bug: 381129034
Test: manual
Change-Id: Ic5f02575b1efc72d5e848907bd1006ca4cdf2edb