Skip Switch To Baseline when Audio Route is not BT.
This CL resolves a call audio routing bug by skipping the switch back to baseline when when call audio is not currently routed to bluetooth in BluetoothDeviceManager#handleAudioRefactoringServiceDisconnected. This resolves a bug where the audio would be errantly routed to earpeiece when BT is disabled while the call is routed to speaker.

Flag: com.android.server.telecom.flags.skip_baseline_switch_when_route_not_bluetooth
Test: build + manual + atest BluetoothDeviceManagerTest
Fixes: 382118789
Change-Id: I355ffcdf94e034a143a057036562b622370f540e