Support TOGGLE_MUTE in CallAudioRouteController
Add explicit support to handle the TOGGLE_MUTE request. This was found
in a bug where the user was unable to mute the call via the headset.
Upon further investigation, I found that we aren't explicitly supporting
handling the request so it ends up being a no-op.

Bug: 372402585
Test: atest CallAudioRouteControllerTest
Test: Manual verification with usb headset
Flag: EXEMPT bugfix
Change-Id: I750c682a2ea348b2e4bcc03f18a9d561c9338652