DSDA: Resolve swap + 3rd outgoing call restriction
For call sequencing, ensure that sequencing is performed across phone
accounts. Previously, this was being handled in Telephony at the
connection level and QC reported issues with unhold happening before
hold. Now, this is being verified in Telecom to ensure that the
operations are occurring sequentially.

Also, there were issues with being able to place a 3rd outgoing call in
the sequencing scenario. This was also previously being handled in
Telephony where if a foreground + background call existed, Telephony
would disconnect the outgoing call. If we have calls on two different
sims, then this logic wouldn't be triggered allowing a 3rd outgoing call
to be placed. Now, this logic is being handled by Telecom to ensure that
we don't allow the third outgoing call.

Bug: 385088638
Bug: 385067389
Bug: 385171137
Flag: com.android.server.telecom.flags.enable_call_sequencing
Test: Manual verification of swap with MT and MO cases and 3rd outgoing
call restricted.
Test: atest TelecomUnitTests
Test: atest CtsTelecomCujTestCases:CallSequencingTests

Change-Id: I90a99c6e875b192851fc47aa51c2743ecdd46967