Attempt to create earpiece route on wired headset disconnect.
In cases where the CallAudioRouteController is initialized with a wired headset plugged in, we will skip creating the earpiece route. So when we do disconnect the headset, the current logic tries to retrieve an existing earpiece audio route mapping which doesn't exist. As a result, we end up not supporting earpiece during audio routing. We should attempt to create the route if it doesn't exist.

Bug: 388206422
Change-Id: Ifebb0b7f7dac1abeacb8ea8f0b385f1fd9d9c346
Flag: EXEMPT bugfix
Test: Manual