Resolve call sequencing computation
Ensure that sequencing is calculated on a transactional basis for the
various actions. Previously, if sequencing was required for any given
transaction, sequencing would be applied for all remaining transactions
in that action. For example, if we had a held call on sub1, an active
call on sub2, and an incoming call on sub2, sequencing would be enforced
between the active call and incoming call. If we had to hold the active
call and then answer the incoming call, we would ensure that the hold
was completed first.

This caused issues when answering an incoming call where because we
waited on hold, the answer command was never sent for the call which
caused the ringing call to be put into a waiting state. In the Telephony
layer, this prevented the active call from being held so the call could
not be answered.

Bug: 387946808
Test: Manual to verify that 3rd incoming call can be answered
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: I3b8b7f7e6d6d51f6ac412698ca958661dab0a26a