DSDA: Resolve autodisconnect + outgoing call error
This CL resolves the auto-disconnect logic in unhold to ensure that
instead of disconnecting the active call, that we fail the transaction
and prompt the user to disconnect the call themselves when we're unable
to hold the active call. This also addresses an issue with not showing
the max outgoing call error dialog when placing a 3rd call when the
target phone account for the new call isn't set. This is possible when
the user switches to DSDS mode when there are two ongoing calls and
tries placing a 3rd call. The accounts are removed due to the
simultaneous calling restrictions and we ended up early exiting from
makeRoomForOutgoingCall with true and displayed a different (irrelevant)
error message instead.

Bug: 390136832
Bug: 388924987
Test: atest TelecomUnitTests
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: I79bc4c6e7c97c9534212bc551a4df368df0f133c