Log BluetoothAdapter#setActiveDevice calls from Telecom.
In a previous cleanup we added logging of calls out from Telecom to
BluetoothHeadset#connectAudio and AudioManager#set/ClearCommunicationDevice.
This was done to make it more clear from the standpoint of Telecom's logs
which audio and BT APIs we were calling.

In a recent bug, I noticed we don't log BluetoothAdapter#setActiveDevice;
this CL adds that.

Test: Manual test - logs only.
Flag: NONE - logs only change.
Fixes: 391417686
Change-Id: I9565f1f7fa5c380801600dd8997b65f6fb928b2a