Use ID 12 for WORK_HANDLE User ID in CallsManagerTest.
WORK_HANDLE was being initialized to User ID 10 in CallsManagerTest. This was causing some of the tests to fail on HSUM devices since on those devices there is an actual user with user ID 10. This CL uses user ID 12 instead to prevent confliction with the actual user 10 on HSUM devices.

Test: CallsManagerTest
Bug: 383203713
Flag: TEST_ONLY
Change-Id: I3a4b3cf9b9c27d58a308432379683fceed1e92c8