Use ID 12 for other user ID in CallsManagerTest.
The content URI of a different user was using  User ID 10 in BasicCallTests. This was causing some of the tests to fail on HSUM devices since on those devices there is an actual user with user ID 10. This CL uses user ID 12 instead to prevent confliction with the actual user 10 on HSUM devices.

Test: BasicCallTests
Bug: 383204897
Flag: TEST_ONLY
Change-Id: Ib0f0443a80b8680c9ad87d7195257c02732b1bcb