Use ID 12 for other user ID in and query current user in TelecomServiceImplTest.
The content URI of a different user was using  User ID 10 in TelecomServiceImplTest. This was causing some of the tests to fail on HSUM devices since on those devices there is an actual user with user ID 10. This CL uses user ID 12 instead to prevent confliction with the actual user 10 on HSUM devices. The same test case was also hard coding in user ID 0 as the current user id. This CL also updates the test to query the calling user ID and generate a URI with that user ID in case the device is HSUM and thus the calling user can be user 10.

Test: TelecomServiceImplTest#testRegisterPhoneAccountImageIconCrossUser
Fixes: 383078188
Flag: TEST_ONLY
Change-Id: I7a4c885c0eb210c00c6b056167c876043ca71ec4