Use User ID 10 for second user in PhoneAccountRegistrarTest.
When creating a list of users in PhoneAccountRegistrarTest#testCleanUpOrphanAccounts, the user IDs 0 and 1000 were being used. In HSUM mode, the user IDs are 0 and 10 out of the box so using IDs 0 and 1000 was causing this test to fail. I have updated the test to use IDs 0 and 10 instead.

Test: PhoneAccountRegistrarTest#testCleanUpOrphanAccounts
Fixes: 383987595
Flag: TEST_ONLY
Change-Id: If21fc89f754bdaa7db399579b03eb74a5d8d410f