Maybe default speaker on wired headset disconnect
When the wired headset disconnects and the previous routing is speaker,
ensure that we route back into speaker. This has been implemented as
part of the legacy routing. Ignore this special cased routing when we're not in call.

Bug: 376364368
Flag: com.android.server.telecom.flags.default_speaker_on_wired_headset_disconnect
Test: Manual verification
Test: atest CallAudioRouteControllerTest

Change-Id: I9920d767725f018ad456356c723adbc2a503f61d