Prevent setCommunicationDevice when not in call
In BluetoothStateReceiver, we attempt to set the communication device
when we receive a new active BT device. We should check before that
we're in a call before doing so. Otherwise, A2DP can be suspended as SCO
will start.

This issue is only noticeable in the new audio HAL as we're setting
communication device for HFP devices now but we should still prevent
this from happening for other BT profiles as well.

Bug: 394880586
Flag: EXEMPT bugfix
Test: Manual
Change-Id: I0f054388aba5a6e3311df7f3ca6b5b1182f1e0f5